requirements.txt
google-api-python-client==2.66.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
python-telegram-bot==20.0.a.6
